agno-building-ai-agents/agno-workflows/pyproject.toml
[project]
name = "agno-workflows"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.2.1",
    "ddgs>=9.6.1",
    "fastapi>=0.120.0",
    "googlesearch-python>=1.3.0",
    "openai>=2.6.1",
    "pycountry>=24.6.1",
    "python-dotenv>=1.1.1",
]
